Module Tezos_shell_benchmarks

module Bloomer_benchmarks : sig ... end

module Encoding_benchmarks : sig ... end

module Encoding_benchmarks_helpers : sig ... end

module Io_benchmarks : sig ... end

module Io_helpers : sig ... end

Helpers for loading contexts, saving contexts, writing to contexts, etc. Also contains the Key_map module, heavily used for preparing benchmarks and computing statistics.

module Io_stats : sig ... end

module Micheline_benchmarks : sig ... end

module Misc_benchmarks : sig ... end