Module Tezos_shell_benchmarks.Io_helpers

Helpers for loading contexts, saving contexts, writing to contexts, etc. Also contains the Key_map module, heavily used for preparing benchmarks and computing statistics.

val assert_ok : msg:string -> ('a, Tezos_error_monad.TzCore.error list) Stdlib.result -> 'a

val prepare_genesis : string -> (Tezos_context.Context.index * Tezos_protocol_environment.Context.t * Tezos_crypto.Context_hash.t, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val commit : Tezos_protocol_environment.Context.t -> Tezos_crypto.Context_hash.t Lwt.t

val prepare_empty_context : string -> (Tezos_crypto.Context_hash.t, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val load_context_from_disk_lwt : string -> Tezos_crypto.Context_hash.t -> (Tezos_protocol_environment.Context.t * Tezos_context.Context.index) Lwt.t

val load_context_from_disk : string -> Tezos_crypto.Context_hash.t -> Tezos_protocol_environment.Context.t * Tezos_context.Context.index

val with_context : base_dir:string -> context_hash:Tezos_crypto.Context_hash.t -> (Tezos_protocol_environment.Context.t -> 'a Lwt.t) -> 'a

val prepare_base_dir : string -> unit

val initialize_key : Stdlib.Random.State.t -> Tezos_protocol_environment.Context.t -> Tezos_protocol_environment.Context.key -> int -> Tezos_protocol_environment.Context.t Lwt.t

val commit_and_reload : string -> Tezos_context.Context.index -> Tezos_protocol_environment.Context.t -> (Tezos_protocol_environment.Context.t * Tezos_context.Context.index) Lwt.t

module Key_map : sig ... end

Maps from string lists to bytes. No balancing. A key cannot be a prefix or a suffix to another key.

val take_n : int -> 'a list -> 'a list -> 'a list * 'a list

val sample_without_replacement : int -> 'a list -> 'a list * 'a list

val buffer_size : int

val buffer : bytes

val file_copy : string -> string -> unit

val set_infos : string -> Unix.stats -> unit

val iter_dir : (string -> unit) -> string -> unit

val copy_rec : string -> string -> unit