Module Tezos_shell_benchmarks__Io_helpers

Helpers for loading contexts, saving contexts, writing to contexts, etc. Also contains the Key_map module, heavily used for preparing benchmarks and computing statistics.